refactor(backend-python): moved logic for requesting database sessions to session.py
