+++ backend-python/app/db/database.py
@@ -1,5 +1,3 @@
-from typing import Generator
-
 from sqlalchemy import Engine, create_engine
 from sqlalchemy.orm import sessionmaker, Session
 
@@ -14,7 +12,9 @@ def setup_database() -> None:
     """
     Initializes database engine and session factory.
 
-    Must be called once at application startup before any call to get_db().
+    This must be called once during application startup BEFORE any call to get_db().
+
+    Without this, SessionLocal will be None and database access will fail.
     """
     global engine, SessionLocal
 
@@ -33,22 +33,3 @@ def setup_database() -> None:
         autoflush=False, 
         bind=engine
     )
-
-
-def get_db() -> Generator[Session, None, None]:
-    """
-    Creates new sessions per request and performs cleanup.
-
-    Requires setup_database() to have been called first.
-
-    Generator is the proper return type hint because of workflow:
-        run → pause (yield) → resume → cleanup
-    """
-    if SessionLocal is None:
-        raise RuntimeError("Database not initialized. Call setup_database() first.")
-
-    db_session = SessionLocal()
-    try:
-        yield db_session
-    finally:
-        db_session.close()
